# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: 面个蛋 MianGeDan API
  version: "1.0.0"
  description: |
    面个蛋（MianGeDan）业务 API v1。追踪 PRD-001 V1.0（docs/prd/PRD-001-面个蛋-V1.0.md）。

    ## 统一约定
    - **幂等**：所有 POST/PATCH/DELETE 写操作必须携带 `Idempotency-Key` 请求头；
      相同键重复提交返回首个结果，不产生重复副作用（NFR-006）。
    - **分页**：列表端点使用 `cursor` + `limit`（默认 20，最大 100），响应含 `next_cursor`。
    - **区域**：`data_region ∈ {cn, eu, intl}` 随 JWT claim 与响应返回；跨区域访问返回 `region_mismatch`。
    - **错误**：统一 Error 结构（见 components.schemas.Error）。
    - **异步任务**：删除、导出、评分、退款返回任务 ID，通过对应 GET 端点查询真实进度。
    - **版本**：路径版本 `/v1`；响应不携带供应商密钥、内部模型端点等机密信息。
servers:
  - url: https://{region}.api.miangedan.example
    description: 区域化 API 入口（三数据区相互隔离）
    variables:
      region:
        enum: [cn, eu, intl]
        default: intl
        description: 数据区域；用户请求必须路由到其账户所属区域
...
paths:
...
  /v1/identity/email/challenges:
    post:
      tags: [identity]
      operationId: createEmailIdentityChallenge
      summary: 创建邮箱验证码挑战（有效期、频率限制与风险验证；不泄露账户是否存在）
      description: |
        邮件必须经 `services/notify` 的所属数据区通道投递。服务端仅持久化邮箱主体摘要与
        验证码散列；通知模板变量不得携带简历/回答正文、业务令牌或媒体。
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, data_region]
              properties:
                email: { type: string, format: email, example: "synthetic.user@example.com" }
                data_region: { $ref: '#/components/schemas/Region' }
      responses:
        '202': { description: 挑战已创建或返回同一幂等结果, content: { application/json: { schema: { $ref: '#/components/schemas/IdentityVerificationChallenge' } } } }
        '429': { $ref: '#/components/responses/RateLimited' }
        default: { $ref: '#/components/responses/Error' }
  /v1/identity/email/challenges/{challengeId}/verify:
    post:
      tags: [identity]
      operationId: verifyEmailIdentityChallenge
      summary: 校验邮箱验证码并签发短期、单次使用的身份验证证明
      security: []
      parameters:
        - $ref: '#/components/parameters/IdentityChallengeId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, data_region]
              properties:
                code: { type: string, pattern: '^[0-9]{6}$' }
                data_region: { $ref: '#/components/schemas/Region' }
      responses:
        '200': { description: 验证成功；证明可用于登录或绑定且只能消费一次, content: { application/json: { schema: { $ref: '#/components/schemas/IdentityVerificationProof' } } } }
        '401': { description: 验证码无效、过期、已消费或超过尝试次数, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        default: { $ref: '#/components/responses/Error' }
  /v1/identity/oauth/{provider}/verify:
    post:
      tags: [identity]
      operationId: verifyOAuthIdentity
      summary: 验证 Google / Apple / 微信身份并签发短期、单次使用证明
      description: |
        调用供应商中立的身份适配器；Google/Apple 仅 eu/intl，微信仅 cn。供应商故障返回
        `provider_unavailable` 且明确 `email_fallback_available=true`，不接收或暴露供应商密钥。
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          schema: { enum: [google, apple, wechat] }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [authorization_code, data_region]
              properties:
                authorization_code: { type: string, minLength: 1, description: 一次性授权码（仅服务端适配器消费） }
                redirect_uri: { type: ['string', 'null'], format: uri }
                data_region: { $ref: '#/components/schemas/Region' }
      responses:
        '200': { description: 验证成功, content: { application/json: { schema: { $ref: '#/components/schemas/IdentityVerificationProof' } } } }
        '503': { description: 第三方不可用；可改用邮箱验证码, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        default: { $ref: '#/components/responses/Error' }
  /v1/identity/sessions:
    post:
      tags: [identity]
      operationId: createIdentitySession
      summary: 使用已验证证明登录；首次身份同时创建所属数据区账户
      description: 未满 16 岁不得创建账户；首次注册必须提交服务条款、隐私政策与数据处理说明版本证据。
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [proof_token, data_region]
              properties:
                proof_token: { type: string, minLength: 32, description: 短期单次身份验证证明；禁止写日志 }
                data_region: { $ref: '#/components/schemas/Region' }
                registration: { oneOf: [{ $ref: '#/components/schemas/IdentityRegistration' }, { type: 'null' }] }
      responses:
        '201': { description: 会话已创建；重复幂等请求返回同一结果, content: { application/json: { schema: { $ref: '#/components/schemas/IdentitySession' } } } }
        '409': { description: 已验证身份与请求区域冲突或注册资料缺失, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        default: { $ref: '#/components/responses/Error' }
  /v1/identity/sessions/refresh:
    post:
      tags: [identity]
      operationId: refreshIdentitySession
      summary: 轮换刷新令牌并签发新业务令牌（旧刷新令牌立即失效）
      security: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token, data_region]
              properties:
                refresh_token: { type: string, minLength: 32, description: 禁止写日志 }
                data_region: { $ref: '#/components/schemas/Region' }
      responses:
        '200': { description: 令牌已轮换, content: { application/json: { schema: { $ref: '#/components/schemas/IdentitySession' } } } }
        '401': { description: 刷新令牌无效、已轮换或已过期, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        default: { $ref: '#/components/responses/Error' }
  /v1/identity/account:
    get:
      tags: [identity]
      operationId: getIdentityAccount
      summary: 获取当前账户与已绑定身份（不返回 provider_subject）
      responses:
        '200': { description: 账户信息, content: { application/json: { schema: { $ref: '#/components/schemas/Account' } } } }
        default: { $ref: '#/components/responses/Error' }
    patch:
      tags: [identity]
      operationId: updateIdentityAccount
      summary: 更新界面语言等账户偏好（界面语言与面试语言相互独立）
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                ui_language: { $ref: '#/components/schemas/Language' }
                display_name: { type: ['string', 'null'], maxLength: 120, description: 可选称呼 }
      responses:
        '200': { description: 已更新, content: { application/json: { schema: { $ref: '#/components/schemas/Account' } } } }
        default: { $ref: '#/components/responses/Error' }
  /v1/identity/bindings:
    post:
      tags: [identity]
      operationId: bindVerifiedIdentity
      summary: 使用当前侧与目标侧两个独立验证证明绑定身份；任何冲突均不合并账户
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source_proof_token, target_proof_token]
              properties:
                source_proof_token: { type: string, minLength: 32, description: 当前账户已绑定身份的重新验证证明 }
                target_proof_token: { type: string, minLength: 32, description: 待绑定身份的独立验证证明 }
      responses:
        '201': { description: 绑定成功；两份证明均被单次消费, content: { application/json: { schema: { $ref: '#/components/schemas/IdentityBinding' } } } }
        '409': { description: 双侧验证不完整或目标身份属于另一账户；未合并，返回恢复案件与人工支持路径, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        default: { $ref: '#/components/responses/Error' }

  # ============================== account data rights ==============================
...
  /v1/projects/{projectId}/rounds/{sequence}/session:
    post:
      tags: [sessions]
      operationId: createSession
      summary: 创建正式/重试轮次房间并签发短期房间令牌（与业务令牌隔离）
      description: |
        前置条件（全部满足，否则 state_conflict/insufficient_entitlement）：
        覆盖方案与量表就绪；第 2 轮起交接包就绪；额度预留成功；单活动设备；便利设置已冻结。
        房间令牌短期有效，仅用于媒体面；浏览器不持有任何供应商密钥。
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/RoundSequence'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, device_id]
              properties:
                kind: { enum: [formal, formal_retry] }
                attempt_id: { type: ['string', 'null'], format: uuid, description: formal_retry 时必填 }
                device_id: { type: string, description: 单活动设备标识 }
      responses:
        '201': { description: 房间已创建, content: { application/json: { schema: { $ref: '#/components/schemas/SessionCreated' } } } }
        '402': { description: 额度不足（阻止开始并提供购买；面试中不提示余额）, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: 前置条件未满足或他设备活动（state_conflict）, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 业务 JWT，含 user_id 与 data_region claim；与媒体房间令牌相互隔离
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 幂等键（UUID 推荐）；相同键重复提交返回首个结果
      schema: { type: string, minLength: 8, maxLength: 128 }
    IdentityChallengeId:
      name: challengeId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ProjectId: { name: projectId, in: path, required: true, schema: { type: string, format: uuid } }
...
    RoundSequence: { name: sequence, in: path, required: true, schema: { type: integer, minimum: 1, maximum: 5 } }
...
  responses:
    Error:
      description: 统一错误响应
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: 触发频率限制
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Region:
      enum: [cn, eu, intl]
      description: 数据区域（三数据区相互隔离，不为容灾擅自跨境）
    Language:
      enum: [zh-CN, en-US]
      description: 首发语言；界面语言与面试语言相互独立
...
    ErrorCode:
      enum: [unauthorized, forbidden, not_found, conflict, validation_failed, upload_rejected, scan_retryable, parse_retryable, sensitive_content_rejected, low_confidence_unresolved, idempotency_conflict, rate_limited, risk_verification_required, verification_invalid, verification_expired, identity_conflict, provider_unavailable, region_mismatch, insufficient_entitlement, state_conflict, internal]
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, trace_id, data_region]
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message: { type: string, description: 用户可见错误（说明影响、数据是否保留、可重试动作、是否计费、是否影响评分） }
            details: { type: ['object', 'null'], additionalProperties: true }
            trace_id: { type: string }
            data_region: { $ref: '#/components/schemas/Region' }
...
    IdentityProvider:
      enum: [email_otp, google, apple, wechat]
      description: 登录身份提供商；手机号码不是必填身份
    IdentityVerificationChallenge:
      type: object
      required: [challenge_id, expires_at, delivery_status, data_region]
      properties:
        challenge_id: { type: string, format: uuid }
        expires_at: { type: string, format: date-time }
        retry_after_seconds: { type: integer, minimum: 0 }
        delivery_status: { enum: [accepted] }
        data_region: { $ref: '#/components/schemas/Region' }
    IdentityVerificationProof:
      type: object
      required: [proof_token, provider, expires_at, data_region]
      properties:
        proof_token: { type: string, minLength: 32, description: 短期单次证明；不得写日志或持久化明文 }
        provider: { $ref: '#/components/schemas/IdentityProvider' }
        expires_at: { type: string, format: date-time }
        data_region: { $ref: '#/components/schemas/Region' }
    IdentityRegistration:
      type: object
      required: [ui_language, age_status, terms_version, privacy_version, data_processing_version, accepted_at, acceptance_context]
      properties:
        ui_language: { $ref: '#/components/schemas/Language' }
        age_status: { enum: [adult, minor_guardian_verified, minor_pending], description: 未满 16 岁仅可使用无登录样例演示 }
        terms_version: { type: string, minLength: 1 }
        privacy_version: { type: string, minLength: 1 }
        data_processing_version: { type: string, minLength: 1 }
        accepted_at: { type: string, format: date-time }
        acceptance_context:
          type: object
          required: [ui_language, surface]
          properties:
            ui_language: { $ref: '#/components/schemas/Language' }
            surface: { enum: [web, pwa, mobile_web] }
          additionalProperties: false
      additionalProperties: false
    IdentityBinding:
      type: object
      required: [identity_id, provider, verified_at, data_region]
      properties:
        identity_id: { type: string, format: uuid }
        provider: { $ref: '#/components/schemas/IdentityProvider' }
        verified_at: { type: string, format: date-time }
        data_region: { $ref: '#/components/schemas/Region' }
    IdentitySession:
      type: object
      required: [session_id, access_token, refresh_token, token_type, expires_in, refresh_expires_in, account, data_region]
      properties:
        session_id: { type: string, format: uuid }
        access_token: { type: string, description: 业务 JWT；禁止写日志 }
        refresh_token: { type: string, description: 单次轮换刷新令牌；禁止写日志 }
        token_type: { const: Bearer }
        expires_in: { type: integer, minimum: 1 }
        refresh_expires_in: { type: integer, minimum: 1 }
        account: { $ref: '#/components/schemas/Account' }
        data_region: { $ref: '#/components/schemas/Region' }
    Account:
      type: object
      required: [user_id, data_region, ui_language, age_status, status]
      properties:
        user_id: { type: string, format: uuid }
        data_region: { $ref: '#/components/schemas/Region' }
        ui_language: { $ref: '#/components/schemas/Language' }
        age_status: { enum: [adult, minor_guardian_verified, minor_pending], description: 未满 16 岁在创建账户前即拒绝，仅可无登录样例演示 }
        status: { enum: [active, deletion_pending, deleted_anonymized] }
        display_name: { type: ['string', 'null'] }
        identities:
          type: array
          items: { $ref: '#/components/schemas/IdentityBinding' }
...
    SessionCreated:
      type: object
      required: [session_id, room_token, room_token_expires_at, data_region]
      properties:
        session_id: { type: string, format: uuid }
        room_url: { type: string, format: uri }
        room_token: { type: string, description: 短期房间令牌（媒体面），与业务令牌相互隔离 }
        room_token_expires_at: { type: string, format: date-time }
        data_region: { $ref: '#/components/schemas/Region' }
...
security:
  - bearerAuth: []